add : unit test and fix some remove kurs code

+++ src/kurs/kurs.service.js
@@ -2,30 +2,28 @@ const db = require('../configs/db.config');
 const { v4: uuidv4 } = require('uuid');
 
 const removeKurs = async (date) => {
-    try {
-        // Delete records from ExchangeRate table
-        const deleteExchangeRateQuery = 'DELETE FROM ExchangeRate WHERE DATE_FORMAT(createdDate, "%Y-%m-%d") = ?';
-        await db.query(deleteExchangeRateQuery, [date]);
+    // Delete records from ExchangeRate table
+    const deleteExchangeRateQuery = 'DELETE FROM ExchangeRate WHERE DATE_FORMAT(createdDate, "%Y-%m-%d") = ?';
+    const [exchangeRateResult] = await db.query(deleteExchangeRateQuery, [date]);
+    const affectedRows = exchangeRateResult.affectedRows;
 
-        // Delete related records from ERate table
-        const deleteERateQuery = 'DELETE FROM ERate WHERE erate_id NOT IN (SELECT erate_id FROM ExchangeRate)';
-        await db.query(deleteERateQuery);
-
-        // Delete related records from TTCounter table
-        const deleteTTCounterQuery = 'DELETE FROM TTCounter WHERE tt_counter_id NOT IN (SELECT tt_counter_id FROM ExchangeRate)';
-        await db.query(deleteTTCounterQuery);
-
-        // Delete related records from BankNotes table
-        const deleteBankNotesQuery = 'DELETE FROM BankNotes WHERE bank_notes_id NOT IN (SELECT bank_notes_id FROM ExchangeRate)';
-        await db.query(deleteBankNotesQuery);
-
-        return { message: 'Records deleted successfully.' };
-
-    } catch (error) {
-        console.error(error);
-        throw new Error('An error occurred while deleting the records.');
+    if (affectedRows === 0) {
+        throw new Error('No kurs found for the specified date.');
     }
 
+    // Delete related records from ERate table
+    const deleteERateQuery = 'DELETE FROM ERate WHERE erate_id NOT IN (SELECT erate_id FROM ExchangeRate)';
+    await db.query(deleteERateQuery);
+
+    // Delete related records from TTCounter table
+    const deleteTTCounterQuery = 'DELETE FROM TTCounter WHERE tt_counter_id NOT IN (SELECT tt_counter_id FROM ExchangeRate)';
+    await db.query(deleteTTCounterQuery);
+
+    // Delete related records from BankNotes table
+    const deleteBankNotesQuery = 'DELETE FROM BankNotes WHERE bank_notes_id NOT IN (SELECT bank_notes_id FROM ExchangeRate)';
+    await db.query(deleteBankNotesQuery);
+
+    return { message: 'Records deleted successfully.' };
 }
 
 const getKurs = async (startdate, enddate) => {
